# Workaround TSDB and timestamp millis precision in mapping