--- a/loadgen/cmd/otelbench/.chloggen/loadgenreceiver-tsdb-millis.yaml
+++ b/loadgen/cmd/otelbench/.chloggen/loadgenreceiver-tsdb-millis.yaml
@@ -1,0 +1,18 @@
+# Use this changelog template to create an entry for release notes.
+
+# One of 'breaking', 'deprecation', 'enhancement', 'bug_fix'
+change_type: enhancement
+
+# A brief description of the change.  Surround your text with quotes ("") if it needs to start with a backtick (`).
+note: Workaround TSDB and timestamp millis precision in mapping
+
+# It is mandatory to specify the component. Do not change this.
+component: loadgenreceiver
+
+# Mandatory: One or more tracking issues related to the change. You can use the PR number here if no issue exists.
+issues: [493]
+
+# (Optional) One or more lines of additional information to render under the primary note.
+# These lines will be padded with 2 spaces and then inserted directly into the document.
+# Use pipe (|) for multiline entries.
+subtext: Add a `original_timestamp` attribute such that data points are not seen as duplicates by TSDB even if loadgenreceiver generate the same metric in the same millisecond.