Workaround: If loadgenreceiver generate data points too quickly,
it is possible that 2 data points of the same metric have timestamps of different nanoseconds (precision sent by elasticsearchexporter)
but the same millisecond (precision in ES index mapping).
This is a problem as it will cause TSDB version_conflict_engine_exception.
Set an original timestamp dimension to ensure that these data points have different dimensions
and will not be seen as duplicates by TSDB.

This removes the need of transformprocessor workaround
```
processors:
  transform/rewrite:
    metric_statements:
      - context: resource
        statements:
          # loadgen may generate datapoints too quickly that multiple datapoints of the same dimensions exist in the same millisecond,
          # which will be seen as duplicates by TSDB
          # Workaround it by adding a random dimension.
          - set(attributes["random"], Substring(MD5(UUID()), 0, 32))
```


Testing: Confirmed that there is no more version_conflict_engine_exception with this PR.