@carsonip 2025-03-27
I doubt it, unless
- the version conflict errors are from a project running otelsoak; and
- the workaround in otelsoak config we use to add a random dimension is not working (very unlikely)