@carsonip 2025-03-27
With this PR, we can remove the workaround to add a random dimension to metrics in our otelsoak config.