@carsonip 2025-03-28
oops, perf is not great. let me see if there's a quick fix
```
                    │ before.txt  │             after.txt              │
                    │   sec/op    │   sec/op     vs base               │
MetricsGenerator-16   1.666m ± 2%   2.890m ± 5%  +73.49% (p=0.002 n=6)

                    │  before.txt  │              after.txt              │
                    │     B/op     │     B/op      vs base               │
MetricsGenerator-16   1.210Mi ± 0%   2.110Mi ± 0%  +74.38% (p=0.002 n=6)

                    │ before.txt  │             after.txt              │
                    │  allocs/op  │  allocs/op   vs base               │
MetricsGenerator-16   16.20k ± 0%   30.15k ± 0%  +86.13% (p=0.002 n=6)
```