@carsonip 2025-03-28
I took a step back, and as es exporter never merges across pmetric.Metrics, seems that adding a resource attribute has lower overhead and have the same effect. Closing this PR in favor of https://github.com/elastic/opentelemetry-collector-components/pull/498